move vpcconfig-building to subcmds pkg
